added user input in main.py

+++ emailcrawler/main.py
@@ -1,4 +1,5 @@
 import os
+from sys import argv
 
 import tldextract
 from scrapy.crawler import CrawlerProcess
@@ -6,9 +7,22 @@ from scrapy.utils.project import get_project_settings
 
 from spiders.emailspider import EmailSpider
 
+if len(argv) != 2:
+    raise ValueError('Please key in the correct number of inputs')
+
+try:
+    input_number = int(argv[1])
+except:
+    raise ValueError('Please key in an integer value')
+
 try:
     os.remove('items.json')
 except OSError:
+    pass
+
+try:
+    os.remove('scrapedata.db')
+except:
     pass
 
 process = CrawlerProcess(get_project_settings())
@@ -16,8 +30,9 @@ counter = 0
 
 with open('/home/bryan/Desktop/tia_selected_websites.csv') as f:
     for line in f:
-        if counter == 50:
-            break
+        if counter == 0:
+            if counter == input_number:
+                break
         start_url = line.split('\t')[1]
         name = line.split('\t')[0]
